challenges/01-architecture-review/services/payment-gateway/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
pika==1.3.2
requests==2.31.0
python-dotenv==1.0.0
